requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.6
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
protobuf>=3.20.0